Rename impl private modules
