+++ src/aiometer/_concurrency/__init__.py
@@ -1,1 +1,1 @@
-from ._channels import *  # noqa: F401, F403
+from .channels import *  # noqa: F401, F403
