+++ src/aiometer/_impl/__init__.py
@@ -1,6 +1,6 @@
-from ._amap import amap
-from ._run_all import run_all
-from ._run_any import run_any
-from ._run_on_each import run_on_each
+from .amap import amap
+from .run_all import run_all
+from .run_any import run_any
+from .run_on_each import run_on_each
 
 __all__ = ["amap", "run_all", "run_any", "run_on_each"]
